chore: mog unused dependencies with cargo-machete
