+++ bolt-sidecar/Cargo.toml
@@ -15,7 +15,6 @@ tokio = { version = "1", features = ["full"] }
 axum = { version = "0.7", features = ["macros"] }
 tower-http = { version = "0.5.2", features = ["timeout"] }
 axum-extra = "0.9.3"
-warp = "0.3.7"
 futures = "0.3"
 tokio-retry = "0.3.0"
 
@@ -24,21 +23,20 @@ blst = "0.3.12"
 tree_hash = "0.5"
 tree_hash_derive = "0.5"
 secp256k1 = { version = "0.29.0", features = ["rand"] }
-ssz_rs = { git = "https://github.com/ralexstokes/ssz-rs" }
+ssz_rs = { git = "https://github.com/ralexstokes/ssz-rs", rev = "ec3073e" }
 ethereum_ssz = "0.5"
-ethereum_ssz_derive = "0.5"
 
 # alloy
-alloy = { version = "0.6.3", features = [
+alloy = { version = "0.6.4", features = [
   "full",
   "provider-trace-api",
   "rpc-types-beacon",
   "rpc-types-engine",
 ] }
+alloy-rpc-types-engine = { version = "0.6.4", default_features = false, features = ["jwt"] }
 
 # reth
 reth-primitives = { git = "https://github.com/paradigmxyz/reth", version = "1.1.1" }
-reth-rpc-layer = { git = "https://github.com/paradigmxyz/reth", version = "1.1.1" }
 
 reqwest = "0.12"
 ethereum-consensus = { git = "https://github.com/ralexstokes/ethereum-consensus", rev = "cf3c404" }
@@ -50,13 +48,11 @@ lighthouse_eth2_keystore = { package = "eth2_keystore", git = "https://github.co
 lighthouse_bls = { package = "bls", git = "https://github.com/sigp/lighthouse", rev = "a87f19d" }
 
 # types
-partial-mpt = { git = "https://github.com/chainbound/partial-mpt", branch = "feat/alloy" }
 serde = { version = "1.0.197", features = ["derive"] }
 serde_json = "1.0.115"
 parking_lot = "0.12.1"
 async-trait = "0.1.79"
 bytes = "1.6.0"
-lru = "0.12.3"
 hex = "0.4.3"
 
 # utils
@@ -65,7 +61,6 @@ thiserror = "1.0"
 rand = "0.8.5"
 dotenvy = "0.15.7"
 regex = "1.10.5"
-toml = "0.5"
 
 # tracing
 tracing = "0.1.40"
@@ -82,7 +77,10 @@ commit-boost = { git = "https://github.com/Commit-Boost/commit-boost-client", re
 cb-common = { git = "https://github.com/Commit-Boost/commit-boost-client", rev = "45ce8f1" }
 
 [dev-dependencies]
-alloy-node-bindings = "0.6.3"
+alloy-node-bindings = "0.6.4" # must match alloy version
+
+[package.metadata.cargo-machete]
+ignored = ["ethereum_ssz"]
 
 
 [[bin]]
